requirements.txt
Flask
Flask-RESTful
pydantic
GitPython
tqdm